+++ tests/conftest.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 import pytest
 
 from siun.config import SiunConfig, Threshold
@@ -13,18 +15,18 @@ def default_thresholds():
 def default_config(default_thresholds):
     """Provide default config."""
     return SiunConfig(
-        **{
-            "cmd_available": "pacman -Quq",
-            "cache_min_age_minutes": 30,
-            "thresholds": default_thresholds,
-            "criteria": {
-                "available_weight": 1,
-                "critical_pattern": "^archlinux-keyring$|^linux$|^pacman.*$",
-                "critical_weight": 1,
-                "count_threshold": 15,
-                "count_weight": 1,
-                "lastupdate_age_hours": 618,  # 7 days
-                "lastupdate_weight": 1,
-            },
-        }
+        cmd_available="pacman -Quq",
+        cache_min_age_minutes=30,
+        thresholds=default_thresholds,
+        criteria={
+            "available_weight": 1,
+            "critical_pattern": "^archlinux-keyring$|^linux$|^pacman.*$",
+            "critical_weight": 1,
+            "count_threshold": 15,
+            "count_weight": 1,
+            "lastupdate_age_hours": 618,  # 7 days
+            "lastupdate_weight": 1,
+        },
+        custom_format="$status_text: $available_updates",
+        state_file=Path("/tmp/siun-state.json"),  # noqa: S108
     )