+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - Unreleased
+
+### Added
+
+- support for choosing state file path via config
+- unit tests for custom output format
+
+### Changed
+
+- ruff linting settings to ignore some less helpful rules
+
+### Fixed
+
+- default value for thresholds setting
+- various unit tests
+
 ## [1.1.0] - 2024-12-05
 
 ### Added